Add a bloom quality setting
The default level is High, which is equivalent to the previous glow
setting used (bicubic upscaling enabled).

Low keeps glow enabled but disables bicubic upscaling.
Disabled disables glow entirely.

The setting is called "bloom" as it's a more common term in gaming,
even though the Environment parameter is called "glow".