Hide menu BackgroundCache node to fully benefit from async. pipeline compilation
This was an attempt by the game to cache the pipelines ahead of time
for the bullet. However, the title screen does not use SDFGI or VoxelGI,
so the ubershader pipeline pre-compilation is not detecting it should
create these pipelines.

The extra two to four pipelines we're seeing on the first shot are the
new pipelines encountered due to the fact the effect only gets enabled
when starting the game, but the instancing of this bullet ahead of time
is not done there instead.

Merely instancing the bullet scene as hidden fixes this issue immediately,
as hidden surfaces are added to the cache.